requirements.txt
Flask
Flask-Cors
gevent
scrapy 
beautifulsoup4 
googlesearch-python